Add files via upload
Thrilled to share our recent success in leveraging Excel for comprehensive call centre trend analysis!

1. Data Cleaning Magic:
In the first leg of our journey, we delved into the depths of call centre data, wielding Excel to transform chaos into clarity. Blanks were replaced with 0, formatting polished - ensuring our data was in shipshape for the adventure ahead. 

2. Crafting Insights with Data Processing:
The second phase saw the birth of new columns, each a piece of the puzzle required to unveil the story within the numbers. Used if functions. Our data processing prowess paved the way for dynamic barcharts and insightful pie charts, giving life to the trends hidden within the digits. 

3. Unveiling the Story - Data Analysis & Reporting:
Now, the moment of truth!  Using pivot tables the report was prepared.Through meticulous data analysis, we unraveled the secrets, calculated key metrics, and painted a vivid picture of call centre performance. Armed with insights, we embarked on the final leg - building a powerful report that not only speaks data but also tells a compelling story. 

The dashboard we've created is more than just numbers - it's a window into the heartbeat of our call centre, providing a real-time understanding of trends, agent performances, and customer satisfaction.